+++ src/deploy/functions/params.spec.ts
@@ -5,6 +5,8 @@ import * as prompt from "../../prompt";
 import * as params from "./params";
 import * as secretManager from "../../gcp/secretManager";
 import { FirebaseError } from "../../error";
+import * as utils from "../../utils";
+import { logger } from "../../logger";
 
 const expect = chai.expect;
 const fakeConfig = {
@@ -78,21 +80,28 @@ describe("CEL resolution", () => {
 
 describe("resolveParams", () => {
   let input: sinon.SinonStub;
+  let logBulletStub: sinon.SinonStub;
+
+  let loggerInfoStub: sinon.SinonStub;
 
   beforeEach(() => {
     input = sinon.stub(prompt, "input");
+    logBulletStub = sinon.stub(utils, "logBullet");
+    loggerInfoStub = sinon.stub(logger, "info");
   });
 
   afterEach(() => {
     input.restore();
+    logBulletStub.restore();
+    loggerInfoStub.restore();
   });
 
   it("always contains the precanned internal param values", async () => {
     const paramsToResolve: params.Param[] = [];
     const userEnv: Record<string, params.ParamValue> = {};
-    await expect(
-      params.resolveParams(paramsToResolve, fakeConfig, userEnv),
-    ).to.eventually.deep.equal(expectedInternalParams);
+    expect(
+      (await params.resolveParams(paramsToResolve, fakeConfig, userEnv, "default")).paramValues,
+    ).to.deep.equal(expectedInternalParams);
   });
 
   it("can pull a literal value out of the dotenvs", async () => {
@@ -111,9 +120,9 @@ describe("resolveParams", () => {
       bar: new params.ParamValue("24", false, { string: false, number: true, boolean: false }),
       baz: new params.ParamValue("true", false, { string: false, number: false, boolean: true }),
     };
-    await expect(
-      params.resolveParams(paramsToResolve, fakeConfig, userEnv),
-    ).to.eventually.deep.equal(
+    expect(
+      (await params.resolveParams(paramsToResolve, fakeConfig, userEnv, "default")).paramValues,
+    ).to.deep.equal(
       Object.assign(
         {
           foo: new params.ParamValue("bar", false, { string: true, number: false, boolean: false }),
@@ -138,9 +147,9 @@ describe("resolveParams", () => {
         boolean: false,
       }),
     };
-    await expect(
-      params.resolveParams(paramsToResolve, fakeConfig, userEnv),
-    ).to.eventually.deep.equal({
+    expect(
+      (await params.resolveParams(paramsToResolve, fakeConfig, userEnv, "default")).paramValues,
+    ).to.deep.equal({
       DATABASE_URL: new params.ParamValue(fakeConfig.databaseURL, true, {
         string: true,
         boolean: false,
@@ -167,13 +176,16 @@ describe("resolveParams", () => {
   it("does not create the corresponding internal params if database url/storage bucket are not configured", async () => {
     const paramsToResolve: params.Param[] = [];
     const userEnv: Record<string, params.ParamValue> = {};
-    await expect(
-      params.resolveParams(
-        paramsToResolve,
-        { locationId: "", projectId: "foo", storageBucket: "", databaseURL: "" },
-        userEnv,
-      ),
-    ).to.eventually.deep.equal({
+    expect(
+      (
+        await params.resolveParams(
+          paramsToResolve,
+          { locationId: "", projectId: "foo", storageBucket: "", databaseURL: "" },
+          userEnv,
+          "default",
+        )
+      ).paramValues,
+    ).to.deep.equal({
       GCLOUD_PROJECT: expectedInternalParams.GCLOUD_PROJECT,
       PROJECT_ID: expectedInternalParams.PROJECT_ID,
     });
@@ -189,7 +201,9 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("bar");
-    await expect(params.resolveParams(paramsToResolve, fakeConfig, {})).to.eventually.deep.equal(
+    expect(
+      (await params.resolveParams(paramsToResolve, fakeConfig, {}, "default")).paramValues,
+    ).to.deep.equal(
       Object.assign(
         {
           foo: new params.ParamValue("bar", false, { string: true }),
@@ -215,7 +229,7 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("baz");
-    await params.resolveParams(paramsToResolve, fakeConfig, {});
+    await params.resolveParams(paramsToResolve, fakeConfig, {}, "default");
     expect(input.getCall(1).args[0].default).to.eq("baz");
   });
 
@@ -235,7 +249,7 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("baz");
-    await params.resolveParams(paramsToResolve, fakeConfig, {});
+    await params.resolveParams(paramsToResolve, fakeConfig, {}, "default");
     expect(input.getCall(1).args[0].default).to.eq("baz/quox");
   });
 
@@ -261,7 +275,7 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("baz");
-    await params.resolveParams(paramsToResolve, fakeConfig, {});
+    await params.resolveParams(paramsToResolve, fakeConfig, {}, "default");
     expect(input.getCall(0).args[0].default).to.eq("https://foo.firebaseio.com/quox");
     expect(input.getCall(1).args[0].default).to.eq("projectID: foo");
     expect(input.getCall(2).args[0].default).to.eq(
@@ -279,7 +293,8 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("");
-    await expect(params.resolveParams(paramsToResolve, fakeConfig, {})).to.eventually.be.rejected;
+    await expect(params.resolveParams(paramsToResolve, fakeConfig, {}, "default")).to.eventually.be
+      .rejected;
   });
 
   it("errors when the default is a CEL expression that resolves to the wrong type", async () => {
@@ -298,7 +313,8 @@ describe("resolveParams", () => {
       },
     ];
     input.resolves("22");
-    await expect(params.resolveParams(paramsToResolve, fakeConfig, {})).to.eventually.be.rejected;
+    await expect(params.resolveParams(paramsToResolve, fakeConfig, {}, "default")).to.eventually.be
+      .rejected;
   });
 
   it("does not throw in non-interactive mode if secret exists in cloud", async () => {
@@ -308,7 +324,8 @@ describe("resolveParams", () => {
       secretVersion: { versionId: "1", state: "ENABLED", secret: {} as any },
     });
 
-    await expect(params.resolveParams(paramsToResolve, fakeConfig, {}, true)).to.be.fulfilled;
+    await expect(params.resolveParams(paramsToResolve, fakeConfig, {}, "default", true)).to.be
+      .fulfilled;
 
     getSecretMetadataStub.restore();
   });
@@ -319,11 +336,35 @@ describe("resolveParams", () => {
       secret: undefined,
     });
 
-    await expect(params.resolveParams(paramsToResolve, fakeConfig, {}, true)).to.be.rejectedWith(
-      FirebaseError,
-      /In non-interactive mode but have no value for the secret/,
-    );
+    await expect(
+      params.resolveParams(paramsToResolve, fakeConfig, {}, "default", true),
+    ).to.be.rejectedWith(FirebaseError, /In non-interactive mode but have no value for the secret/);
 
     getSecretMetadataStub.restore();
+  });
+
+  it("never calls ensureSecret if running in the emulator", async () => {
+    const paramsToResolve: params.Param[] = [{ name: "MY_SECRET", type: "secret" }];
+    const getSecretMetadataSpy = sinon.spy(secretManager, "getSecretMetadata");
+
+    await params.resolveParams(paramsToResolve, fakeConfig, {}, "default", false, false, true);
+    expect(getSecretMetadataSpy.called).to.be.false;
+
+    getSecretMetadataSpy.restore();
+  });
+
+  it("should print a header when prompting for a codebase", async () => {
+    const paramsToResolve: params.Param[] = [
+      {
+        name: "foo",
+        type: "string",
+        input: { text: {} },
+      },
+    ];
+    input.resolves("bar");
+    await params.resolveParams(paramsToResolve, fakeConfig, {}, "my-codebase");
+    expect(
+      loggerInfoStub.calledWith(sinon.match(/Prompting for parameters for codebase.*my-codebase/)),
+    ).to.be.true;
   });
 });