+++ src/gcp/cloudfunctionsv2.ts
@@ -503,6 +503,21 @@ export function functionFromEndpoint(endpoint: backend.Endpoint): InputCloudFunc
   proto.convertIfPresent(gcfFunction.serviceConfig, endpoint, "availableCpu", "cpu", (cpu) => {
     return String(cpu);
   });
+  // V1 Functions had a an automatic FUNCTION_REGION environment variable. V2 functions omitted this because Cloud Run's original
+  // goal was to create a platform more portable to KNative. We're preserving the old behaviour by adding this back in so that
+  // users who depended on this (e.g. generic code that targeted a task queue in the same region) don't break on upgrade.
+  // We only set it when the user hasn't explicitly set it, since we don't want to override a user's explicit setting.
+  if (!gcfFunction.serviceConfig.environmentVariables) {
+    gcfFunction.serviceConfig.environmentVariables = {};
+  }
+  if (!endpoint.region) {
+    throw new FirebaseError(
+      `Function ${endpoint.id} is missing the function region at deploy time. This should never happen`,
+      { exit: 1 },
+    );
+  }
+  // N.B. This var is reserved so we don't need to check if it's been set already.
+  gcfFunction.serviceConfig.environmentVariables["FUNCTION_REGION"] = endpoint.region;
 
   if (endpoint.vpc) {
     if (endpoint.vpc.connector) {